+++ core/src/main/java/com/backblaze/b2/client/structures/B2UpdateBucketRequest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021, Backblaze Inc. All Rights Reserved.
+ * Copyright 2022, Backblaze Inc. All Rights Reserved.
  * License https://www.backblaze.com/using_b2_code.html
  */
 package com.backblaze.b2.client.structures;
@@ -39,10 +39,13 @@ public class B2UpdateBucketRequest {
     private final B2BucketReplicationConfiguration replicationConfiguration;
 
     @B2Json.optional
+    private final Boolean fileLockEnabled;
+
+    @B2Json.optional
     private final Integer ifRevisionIs;
 
     @B2Json.constructor(params = "accountId,bucketId,bucketType,bucketInfo,corsRules,lifecycleRules," +
-            "defaultRetention,defaultServerSideEncryption,replicationConfiguration,ifRevisionIs")
+            "defaultRetention,defaultServerSideEncryption,replicationConfiguration,fileLockEnabled,ifRevisionIs")
     private B2UpdateBucketRequest(String accountId,
                                   String bucketId,
                                   String bucketType,
@@ -52,6 +55,7 @@ public class B2UpdateBucketRequest {
                                   B2BucketDefaultRetention defaultRetention,
                                   B2BucketServerSideEncryption defaultServerSideEncryption,
                                   B2BucketReplicationConfiguration replicationConfiguration,
+                                  Boolean fileLockEnabled,
                                   Integer ifRevisionIs) {
         this.accountId = accountId;
         this.bucketId = bucketId;
@@ -62,6 +66,7 @@ public class B2UpdateBucketRequest {
         this.defaultRetention = defaultRetention;
         this.defaultServerSideEncryption = defaultServerSideEncryption;
         this.replicationConfiguration = replicationConfiguration;
+        this.fileLockEnabled = fileLockEnabled;
         this.ifRevisionIs = ifRevisionIs;
     }
 
@@ -100,6 +105,10 @@ public class B2UpdateBucketRequest {
         return replicationConfiguration;
     }
 
+    public Boolean getFileLockEnabled() {
+        return fileLockEnabled;
+    }
+
     public Integer getIfRevisionIs() {
         return ifRevisionIs;
     }
@@ -118,6 +127,7 @@ public class B2UpdateBucketRequest {
                 Objects.equals(getDefaultRetention(), that.getDefaultRetention()) &&
                 Objects.equals(getDefaultServerSideEncryption(), that.getDefaultServerSideEncryption()) &&
                 Objects.equals(getReplicationConfiguration(), that.getReplicationConfiguration()) &&
+                Objects.equals(getFileLockEnabled() , that.getFileLockEnabled()) &&
                 Objects.equals(getIfRevisionIs(), that.getIfRevisionIs());
     }
 
@@ -133,6 +143,7 @@ public class B2UpdateBucketRequest {
                 getDefaultRetention(),
                 getDefaultServerSideEncryption(),
                 getReplicationConfiguration(),
+                getFileLockEnabled(),
                 getIfRevisionIs()
         );
     }
@@ -156,6 +167,7 @@ public class B2UpdateBucketRequest {
         private B2BucketDefaultRetention defaultRetention;
         private B2BucketServerSideEncryption defaultServerSideEncryption;
         private B2BucketReplicationConfiguration replicationConfiguration;
+        private Boolean fileLockEnabled;
 
         private Builder(B2Bucket bucket) {
             this.accountId = bucket.getAccountId();
@@ -198,6 +210,11 @@ public class B2UpdateBucketRequest {
             return this;
         }
 
+        public Builder setFileLockEnabled(Boolean fileLockEnabled) {
+            this.fileLockEnabled = fileLockEnabled;
+            return this;
+        }
+
         public B2UpdateBucketRequest build() {
             return new B2UpdateBucketRequest(
                     accountId,
@@ -209,6 +226,7 @@ public class B2UpdateBucketRequest {
                     defaultRetention,
                     defaultServerSideEncryption,
                     replicationConfiguration,
+                    fileLockEnabled,
                     ifRevisionIs
             );
         }